Separate public and internal hooks, reduce exports
JIRA: RAIL-2956
