+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/DashboardView/DashboardView.tsx
@@ -5,12 +5,15 @@ import { ThemeProvider, useThemeIsLoading } from "@gooddata/sdk-ui-theme-provide
 import { isFluidLayoutEmpty } from "@gooddata/sdk-backend-spi";
 import { idRef } from "@gooddata/sdk-model";
 
-import { useDashboard } from "../hooks/useDashboard";
-import { useDashboardAlerts } from "../hooks/useDashboardAlerts";
+import { useDashboard, useDashboardAlerts } from "../hooks";
+import {
+    useAttributesWithDrillDown,
+    useColorPalette,
+    useDashboardViewLayout,
+    useUserWorkspaceSettings,
+} from "../hooks/internal";
 import { IDashboardViewConfig, IDashboardViewProps } from "./types";
 import { InternalIntlWrapper } from "../../utils/internalIntlProvider";
-import { useUserWorkspaceSettings } from "../hooks/useUserWorkspaceSettings";
-import { useColorPalette } from "../hooks/useColorPalette";
 import {
     AttributesWithDrillDownProvider,
     ColorPaletteProvider,
@@ -21,8 +24,6 @@ import {
 } from "./contexts";
 import { defaultThemeModifier } from "./defaultThemeModifier";
 import { ScheduledMailDialog } from "../ScheduledMail/ScheduledMailDialog/ScheduledMailDialog";
-import { useAttributesWithDrillDown } from "../hooks/useAttributesWithDrillDown";
-import { useDashboardViewLayout } from "../hooks/useDashboardViewLayout";
 import { DashboardRenderer } from "./DashboardRenderer";
 import { EmptyDashboardError } from "./EmptyDashboardError";
 import { filterArrayToFilterContextItems } from "../utils/filters";
