+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/DashboardView/KpiView/alertManipulationHooks/useAlertSaveOrUpdateHandler.ts
@@ -1,7 +1,7 @@
 // (C) 2021 GoodData Corporation
 import { useState } from "react";
 import { IAnalyticalBackend, isWidgetAlert, IWidgetAlertDefinition } from "@gooddata/sdk-backend-spi";
-import { useSaveOrUpdateWidgetAlert } from "../../../hooks/useSaveOrUpdateWidgetAlert";
+import { useSaveOrUpdateWidgetAlert } from "../../../hooks/internal";
 import { KpiAlertOperationStatus } from "../../../types";
 import { useAlerts } from "../../contexts";
