+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/hooks/dataLoaders/DateDatasetsDataLoader.ts
@@ -2,10 +2,7 @@
 import { IAnalyticalBackend, ICatalogDateDataset } from "@gooddata/sdk-backend-spi";
 import { dataLoaderAbstractFactory } from "../../../../dataLoaders/DataLoaderAbstractFactory";
 
-/**
- * @internal
- */
-export interface IDateDatasetsDataLoader {
+interface IDateDatasetsDataLoader {
     /**
      * Obtains all catalog attributes with a drill down specified.
      * @param backend - the {@link IAnalyticalBackend} instance to use to communicate with the backend
