+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/hooks/index.ts
@@ -1,0 +1,5 @@
+// (C) 2021 GoodData Corporation
+export * from "./useDashboard";
+export * from "./useDashboardAlerts";
+export * from "./useDashboardPdfExporter";
+export * from "./useDashboardWidgetExecution";
