+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/hooks/internal/index.ts
@@ -1,0 +1,16 @@
+// (C) 2021 GoodData Corporation
+export * from "./useAttributesWithDrillDown";
+export * from "./useColorPalette";
+export * from "./useCurrentUser";
+export * from "./useDashboardInsightExecution";
+export * from "./useDashboardKpiExecution";
+export * from "./useDashboardViewLayout";
+export * from "./useDeleteWidgetAlert";
+export * from "./useExportDashboardToPdf";
+export * from "./useKpiData";
+export * from "./useSaveOrUpdateWidgetAlert";
+export * from "./useSaveScheduledMail";
+export * from "./useUpdateWidgetAlert";
+export * from "./useUserWorkspacePermissions";
+export * from "./useUserWorkspaceSettings";
+export * from "./useWorkspaceUsers";
