+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/hooks/internal/useSaveOrUpdateWidgetAlert.ts
@@ -1,5 +1,10 @@
 // (C) 2020-2021 GoodData Corporation
-import { IAnalyticalBackend, IWorkspacePermissions } from "@gooddata/sdk-backend-spi";
+import {
+    IAnalyticalBackend,
+    isWidgetAlert,
+    IWidgetAlert,
+    IWidgetAlertDefinition,
+} from "@gooddata/sdk-backend-spi";
 import {
     GoodDataSdkError,
     useBackend,
@@ -8,14 +13,15 @@ import {
     UseCancelablePromiseState,
     useWorkspace,
 } from "@gooddata/sdk-ui";
-import { userWorkspacePermissionsDataLoaderFactory } from "./dataLoaders";
-import { backendInvariant, workspaceInvariant } from "./utils";
+import { backendInvariant, workspaceInvariant } from "../utils";
 
-/**
- * @beta
- */
-export interface IUseUserWorkspacePermissionsConfig
-    extends UseCancelablePromiseCallbacks<IWorkspacePermissions, GoodDataSdkError> {
+interface IUseSaveOrUpdateWidgetAlertConfig
+    extends UseCancelablePromiseCallbacks<IWidgetAlert, GoodDataSdkError> {
+    /**
+     * Widget alert to save or update.
+     */
+    widgetAlert?: IWidgetAlertDefinition | IWidgetAlert;
+
     /**
      * Backend to work with.
      *
@@ -34,30 +40,36 @@ export interface IUseUserWorkspacePermissionsConfig
 }
 
 /**
- * Hook allowing to download user workspace permissions
+ * Hook allowing to save a widget alert
  * @param config - configuration of the hook
- * @beta
+ * @internal
  */
-export function useUserWorkspacePermissions({
+export function useSaveOrUpdateWidgetAlert({
+    widgetAlert,
     backend,
+    workspace,
     onCancel,
     onError,
     onLoading,
     onPending,
     onSuccess,
-    workspace,
-}: IUseUserWorkspacePermissionsConfig): UseCancelablePromiseState<IWorkspacePermissions, any> {
+}: IUseSaveOrUpdateWidgetAlertConfig): UseCancelablePromiseState<IWidgetAlert, any> {
     const effectiveBackend = useBackend(backend);
     const effectiveWorkspace = useWorkspace(workspace);
 
-    backendInvariant(effectiveBackend, "useUserWorkspacePermissions");
-    workspaceInvariant(effectiveWorkspace, "useUserWorkspacePermissions");
+    backendInvariant(effectiveBackend, "useSaveOrUpdateWidgetAlert");
+    workspaceInvariant(effectiveWorkspace, "useSaveOrUpdateWidgetAlert");
 
-    const loader = userWorkspacePermissionsDataLoaderFactory.forWorkspace(effectiveWorkspace);
-    const promise = () => loader.getUserWorkspacePermissions(effectiveBackend);
+    const promise = widgetAlert
+        ? () =>
+              isWidgetAlert(widgetAlert)
+                  ? effectiveBackend.workspace(effectiveWorkspace).dashboards().updateWidgetAlert(widgetAlert)
+                  : effectiveBackend.workspace(effectiveWorkspace).dashboards().createWidgetAlert(widgetAlert)
+        : null;
 
     return useCancelablePromise({ promise, onCancel, onError, onLoading, onPending, onSuccess }, [
         effectiveBackend,
         effectiveWorkspace,
+        widgetAlert,
     ]);
 }
