+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/hooks/internal/useUpdateWidgetAlert.ts
@@ -1,5 +1,5 @@
 // (C) 2020-2021 GoodData Corporation
-import { IAnalyticalBackend, IUserWorkspaceSettings } from "@gooddata/sdk-backend-spi";
+import { IAnalyticalBackend, IWidgetAlert } from "@gooddata/sdk-backend-spi";
 import {
     GoodDataSdkError,
     useBackend,
@@ -8,14 +8,14 @@ import {
     UseCancelablePromiseState,
     useWorkspace,
 } from "@gooddata/sdk-ui";
-import { userWorkspaceSettingsDataLoaderFactory } from "../../../dataLoaders";
-import { backendInvariant, workspaceInvariant } from "./utils";
+import { backendInvariant, workspaceInvariant } from "../utils";
 
-/**
- * @beta
- */
-export interface IUseUserWorkspaceSettingsConfig
-    extends UseCancelablePromiseCallbacks<IUserWorkspaceSettings, GoodDataSdkError> {
+interface IUseUpdateWidgetAlertConfig extends UseCancelablePromiseCallbacks<IWidgetAlert, GoodDataSdkError> {
+    /**
+     * Widget alert to update.
+     */
+    widgetAlert?: IWidgetAlert;
+
     /**
      * Backend to work with.
      *
@@ -25,7 +25,7 @@ export interface IUseUserWorkspaceSettingsConfig
     backend?: IAnalyticalBackend;
 
     /**
-     * Workspace where the insight exists.
+     * Workspace to work with.
      *
      * Note: the workspace must come either from this property or from WorkspaceContext. If you do not specify
      * workspace here, then the hook MUST be called within an existing WorkspaceContext.
@@ -34,30 +34,33 @@ export interface IUseUserWorkspaceSettingsConfig
 }
 
 /**
- * Hook allowing to download user workspace settings
+ * Hook allowing to update a widget alert
  * @param config - configuration of the hook
- * @beta
+ * @internal
  */
-export function useUserWorkspaceSettings({
+export function useUpdateWidgetAlert({
+    widgetAlert,
     backend,
+    workspace,
     onCancel,
     onError,
     onLoading,
     onPending,
     onSuccess,
-    workspace,
-}: IUseUserWorkspaceSettingsConfig): UseCancelablePromiseState<IUserWorkspaceSettings, any> {
+}: IUseUpdateWidgetAlertConfig): UseCancelablePromiseState<IWidgetAlert, any> {
     const effectiveBackend = useBackend(backend);
     const effectiveWorkspace = useWorkspace(workspace);
 
-    backendInvariant(effectiveBackend, "useUserWorkspaceSettings");
-    workspaceInvariant(effectiveWorkspace, "useUserWorkspaceSettings");
+    backendInvariant(effectiveBackend, "useUpdateWidgetAlert");
+    workspaceInvariant(effectiveWorkspace, "useUpdateWidgetAlert");
 
-    const loader = userWorkspaceSettingsDataLoaderFactory.forWorkspace(effectiveWorkspace);
-    const promise = () => loader.getUserWorkspaceSettings(effectiveBackend);
+    const promise = widgetAlert
+        ? () => effectiveBackend.workspace(effectiveWorkspace).dashboards().updateWidgetAlert(widgetAlert)
+        : null;
 
     return useCancelablePromise({ promise, onCancel, onError, onLoading, onPending, onSuccess }, [
         effectiveBackend,
         effectiveWorkspace,
+        widgetAlert,
     ]);
 }
