+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/hooks/internal/useUserWorkspaceSettings.ts
@@ -1,6 +1,5 @@
 // (C) 2020-2021 GoodData Corporation
-import { IAnalyticalBackend } from "@gooddata/sdk-backend-spi";
-import { IColorPalette } from "@gooddata/sdk-model";
+import { IAnalyticalBackend, IUserWorkspaceSettings } from "@gooddata/sdk-backend-spi";
 import {
     GoodDataSdkError,
     useBackend,
@@ -9,14 +8,11 @@ import {
     UseCancelablePromiseState,
     useWorkspace,
 } from "@gooddata/sdk-ui";
-import { colorPaletteDataLoaderFactory } from "../../../dataLoaders";
-import { backendInvariant, workspaceInvariant } from "./utils";
+import { userWorkspaceSettingsDataLoaderFactory } from "../../../../dataLoaders";
+import { backendInvariant, workspaceInvariant } from "../utils";
 
-/**
- * @beta
- */
-export interface IUseColorPaletteConfig
-    extends UseCancelablePromiseCallbacks<IColorPalette, GoodDataSdkError> {
+interface IUseUserWorkspaceSettingsConfig
+    extends UseCancelablePromiseCallbacks<IUserWorkspaceSettings, GoodDataSdkError> {
     /**
      * Backend to work with.
      *
@@ -26,7 +22,7 @@ export interface IUseColorPaletteConfig
     backend?: IAnalyticalBackend;
 
     /**
-     * Workspace to get the color palette for.
+     * Workspace where the insight exists.
      *
      * Note: the workspace must come either from this property or from WorkspaceContext. If you do not specify
      * workspace here, then the hook MUST be called within an existing WorkspaceContext.
@@ -35,11 +31,11 @@ export interface IUseColorPaletteConfig
 }
 
 /**
- * Hook allowing to download color palette of the given workspace
+ * Hook allowing to download user workspace settings
  * @param config - configuration of the hook
- * @beta
+ * @internal
  */
-export function useColorPalette({
+export function useUserWorkspaceSettings({
     backend,
     onCancel,
     onError,
@@ -47,15 +43,15 @@ export function useColorPalette({
     onPending,
     onSuccess,
     workspace,
-}: IUseColorPaletteConfig): UseCancelablePromiseState<IColorPalette, any> {
+}: IUseUserWorkspaceSettingsConfig): UseCancelablePromiseState<IUserWorkspaceSettings, any> {
     const effectiveBackend = useBackend(backend);
     const effectiveWorkspace = useWorkspace(workspace);
 
-    backendInvariant(effectiveBackend, "useColorPalette");
-    workspaceInvariant(effectiveWorkspace, "useColorPalette");
+    backendInvariant(effectiveBackend, "useUserWorkspaceSettings");
+    workspaceInvariant(effectiveWorkspace, "useUserWorkspaceSettings");
 
-    const loader = colorPaletteDataLoaderFactory.forWorkspace(effectiveWorkspace);
-    const promise = () => loader.getColorPalette(effectiveBackend);
+    const loader = userWorkspaceSettingsDataLoaderFactory.forWorkspace(effectiveWorkspace);
+    const promise = () => loader.getUserWorkspaceSettings(effectiveBackend);
 
     return useCancelablePromise({ promise, onCancel, onError, onLoading, onPending, onSuccess }, [
         effectiveBackend,
